feat(frontend) Be able to have multiple frontends.
Split the models and MySQL into their own modules.
