+++ src/Frontend/MySQL/Column.php
@@ -1,0 +1,12 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Hywan\DatabaseToPlantUML\Frontend\MySQL;
+
+use Hoa\Visitor;
+use Hywan\DatabaseToPlantUML\Frontend;
+
+class Column extends Frontend\Column
+{
+}
